added hop interval argument
Before now, the hop interval was hard-coded as 5 seconds. This was a reasonable default for scanning two or three channels, but for more channels a smaller hop interval makes sense. To account for more use cases, it seemed reasonable to give the user the option of configuring the hop interval.