sort db chronologically in usage command
The logic in the usage command already assumed the database was sorted. Now, I am actually meeting that assumption. It should not be necessary to supply a sorted database; the database might be a concatenated set of data collected from multiple locations, in which case various times would be overlapping.