created real CSV parser in db_entry.c
Before, I used a silly "dialect" of CSV which used backslashes for escapes. This is not how the actual CSV specification works. Now, I have a true CSV parser. I have to admit that this format is harder to parse than the old one, but it actually results in a *slightly* smaller log file (at least for web responses with dates in them, since they have a bunch of commas).